update test filter function
