+++ cypress/support/commands.js
@@ -63,3 +63,15 @@ Cypress.Commands.add('loginToAuth0', () => {
 });
 
 Cypress.Commands.add('getReduxState', () => cy.window().its('store').invoke('getState'));
+
+Cypress.Commands.add('assertByTestId', (testId) => {
+  cy.get(`[data-test=${testId}]`).should('exist');
+});
+
+Cypress.Commands.add('beforeEachVisitBaseUrl', () => {
+  cy.visit('');
+  // Check for and click the "not now" button if it exists
+  cy.contains(/not now/i).click({ multiple: true, force: true });
+  // Check for and click the "decline" button if it exists
+  cy.contains(/decline/i).click({ multiple: true, force: true });
+});
